Created AppointmentForm component
