+++ src/assets/css/styles.css
@@ -191,3 +191,23 @@ input:focus {
 .input-group button {
   flex: 2 1 auto;
 }
+
+/*
+ * Radio buttons
+ */
+.radio__buttons {
+  border: none;
+  display: grid;
+  grid-template-columns: 1fr 1fr 1fr;
+  column-gap: 1rem;
+  row-gap: 0.5rem;
+}
+
+.radio__buttons > .option {
+  cursor: pointer;
+  padding: 0.5rem 0;
+  white-space: nowrap;
+}
+.radio__buttons > .option.option--selected {
+  color: var(--primary-shade);
+}
